refactored code following PR suggestions
